# swagger.json
{"swagger":"2.0","info":{"title":"NBA21 API Documentation","description":"This is the official documentation of NBA21.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ListUserDto":{"type":"object","properties":{"search":{"type":"string"},"name":{"type":"string"},"names":{"type":"string"}}},"AuthCredentialsDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}},"paths":{"/api/json":{"get":{"operationId":"AppController_getSwagger","summary":"Get Swagger JSON","parameters":[],"responses":{"200":{"description":""}}}},"/user":{"get":{"operationId":"UserController_getUsers","summary":"List Users","description":"List all registered users.","parameters":[],"requestBody":{"required":true,"description":"List all registered users.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"JWT":[]}]}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","summary":"Sign Up","description":"Sign up. supply username & password to create a new user in the system.","parameters":[],"requestBody":{"required":true,"description":"User credentials: username & password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCredentialsDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}},"/auth/signin":{"post":{"operationId":"AuthController_singIn","summary":"Sign In","description":"Sign in. supply username & password and get JWT token in return.","parameters":[],"requestBody":{"required":true,"description":"User credentials: username & password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthCredentialsDto"}}}},"responses":{"200":{"description":"Returns JWT token of this user"},"400":{"description":"Indicates that one (or more) of the given parameters was invalid."},"401":{"description":"Indicates that the given credentials are incorrect."}},"tags":["Authentication"]}},"/auth/test":{"post":{"operationId":"AuthController_test","summary":"Test","description":"Test Token. supply JWT token and get matching username. useful for debug purposes.","parameters":[],"responses":{"201":{"description":""}},"tags":["Authentication"],"security":[{"JWT":[]}]}}}}